requirements.txt
Cython>=0.29
h5py>=2.5.0
numpy>=1.10.4
Pillow>=3.1.1
PyYAML>=3.11
scipy>=0.17.0
scikit-learn>=0.17.0
six>=1.10.0
tensorflow-gpu==2.5
